requirements.txt
google-cloud-aiplatform
google-generativeai
jinja2
weasyprint
tqdm
python-dotenv
banana-dev
requests
pillow
